requirements.txt
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mecab-python3==0.996.2
redis==3.2.1
six==1.12.0
Werkzeug==0.15.4
